requirements.txt
pyats[full]>=24.0
genie>=24.0
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1
urllib3>=2.0
